add playerMute, playerDeaf event | add function disconnect Node (only ws)

+++ src/structures/NodeManager.ts
@@ -1,6 +1,6 @@
 import { EventEmitter } from "events";
 
-import { DestroyReasons } from "./Constants";
+import { DestroyReasons, DisconnectReasons } from "./Constants";
 import { LavalinkNode } from "./Node";
 import { MiniMap } from "./Utils";
 
@@ -84,13 +84,17 @@ export class NodeManager extends EventEmitter {
      * @param deleteAllNodes if the nodes should also be deleted from nodeManager.nodes
      * @returns amount of disconnected Nodes
      */
-    public async disconnectAll(deleteAllNodes = false) {
+    public async disconnectAll(deleteAllNodes = false, destroyPlayers = true) {
         if (!this.nodes.size) throw new Error("There are no nodes to disconnect (no nodes in the nodemanager)");
         if (!this.nodes.filter(v => v.connected).size) throw new Error("There are no nodes to disconnect (all nodes disconnected)");
         let counter = 0;
         for (const node of [...this.nodes.values()]) {
             if (!node.connected) continue;
-            await node.destroy(DestroyReasons.DisconnectAllNodes, deleteAllNodes);
+            if (destroyPlayers) {
+                await node.destroy(DestroyReasons.DisconnectAllNodes, deleteAllNodes);
+            } else {
+                await node.disconnect(DisconnectReasons.DisconnectAllNodes);
+            }
             counter++;
         }
         return counter;
